add: move label tool bbox to module file
